requirements.txt
python-telegram-bot
python-dotenv
openai
asyncpg
langchain
langchain-openai
langchain-community
sqlalchemy
psycopg2-binary
